rename: components 디렉터리 내 파일 이동 (components/cartItem 디렉터리 생성)
